添加权限

+++ cqytjgpt-webapi/cqytjgpt-webapi-repository/src/main/java/com/ccttic/cqytjgpt/webapi/mapper/danger/DrDangerMapper.java
@@ -63,6 +63,6 @@ public interface DrDangerMapper {
 	 * @param params
 	 * @return
 	 */
-	long qryDriverListCount(Map<String, Object> params);
+	int qryDriverListCount(Map<String, Object> params);
 
 }
